requirements.txt
--extra-index-url https://devpi.iast.it/nephila/public/+simple/
Django==1.9.12
Markdown==2.6
django-bootstrap3==7.0.0
django-filter==0.9.2
django-autoslug==1.9.3
django-mptt==0.8.6
phonenumbers==7.0.2
Pillow==3.2.0
django-countries==3.3
googlemaps==2.2
lxml==3.5.0
psycopg2==2.6.1
names==0.3.0
django-autocomplete-light==2.3.3
django-model-utils==2.6
django-ckeditor==5.0.3
XlsxWriter==0.7.7
django-extensions==1.5.9
django-multiupload==0.5
ftfy==4.0
django-loginas==0.1.8
easy-thumbnails==2.3
python-stdnum==1.2
codicefiscale==0.9
django-cron==0.4.6
django-gravatar2==1.4.0
newrelic==2.60.0.46
django_compressor==2.0
csscompressor==0.9.4
slimit==0.8.1
pyBarcode==0.7
cssselect==0.9.1
splinter==0.7.5
tblib==1.3.0
selenium==2.53.6
django-filer==1.2.5.post1
django-ckeditor-filebrowser-filer==0.2.0
django-nocaptcha-recaptcha==0.0.19
django-two-factor-auth==1.4.0
django-otp-yubikey==0.3.4
freezegun
unicodecsv==0.14.1